requirements.txt
pyttsx3==2.90
transformers==4.44.2
sentencepiece==0.2.0
accelerate==0.34.2
vosk==0.3.45
rich==13.8.0
requests==2.32.3
python-dotenv==1.0.1
gtts==2.5.1
playsound==1.3.0